chore: Add WallRemove event and handle it in Map::InitEvent()
This commit adds a new event type, WallRemove, to the EventType enum in EventType.h. It also introduces a new event class, WallRemoveEventArgs, in EventArgs.h, which inherits from EventArgs and includes a GetMapIndex() method. In Map::InitEvent() in Event.cpp, a new listener is appended to the m_Event object to handle the WallRemove event. The listener calls the RemoveWall() method with the map index provided in the WallRemoveEventArgs. This change allows for the removal of walls in the game when the corresponding event is triggered.