feat: Add FloorUpdate event and handle it in Map::TempoTrigger()
This commit adds a new event type, FloorUpdate, to the EventType enum in EventType.h. It also introduces a new event class, FloorUpdateEventArgs, in EventArgs.h, which inherits from EventArgs and includes a GetTempoIndex() and GetColored() method. In Map::TempoTrigger() in Map.cpp, the FloorUpdate event is dispatched with the corresponding arguments. This change allows for the updating of the floor in the game when the tempo changes, and also provides information about the tempo index and whether the floor should be colored.