chore: Refactor MapData.h and ClickEvent.cpp
This commit refactors the MapData.h and ClickEvent.cpp files. In MapData.h, the m_Player member variable is now initialized as nullptr. In ClickEvent.cpp, the code for loading a new level has been modified to only execute if the loadLevel variable is true. This change ensures that the music, tempo, and animation destinations are only updated when a new level is successfully loaded.