feat: Add flag to Music::IndicatorBar for correct display
This commit adds a flag to the Music::IndicatorBar class to ensure the correct display of the indicator bar. The flag is used to control the animation and position of the left and right indicators based on the current beat index and tempo number. This change improves the accuracy of the indicator bar's display in the game.